Remove explicit namespaces from Helm charts in favour of release namespace
When installing a Helm chart, you install it to a specific namespace. Resources that don't already specify a namespace will
be installed into the same namespace as the rest of the chart. Explicitly defining namespaces, or defining them
in Values.yaml will override this. That is usually less portable, unless a chart needs to create or install into multiple namespaces.

Fix HS-274