Add partially-working Tiltfile
This adds a Tiltfile that almost works. It uses k8s_attach to get the pods created by the operator, but file sync
doesn't fully work, the same logs are streamed to 'opennms-crd' and the actual pod/build resources, and build/runtime
logs are not present in the same resources.

Fix HS-276